+++ packages/terra-menu/src/Menu.jsx
@@ -55,12 +55,17 @@ const propTypes = {
    */
   isArrowDisplayed: PropTypes.bool,
   /**
+   * ![IMPORTANT](https://badgen.net/badge/prop/Important/red)
    * Header Title for main-menu(first-tier).
    * Header Title will only be visible if the main-menu contains at least one sub-menu.
+   * `headerTitle` is expected to be provided as it is used in cases of responsive
+   * behavior when Menu is bounded/fullscreen.
    */
   headerTitle: PropTypes.string,
   /**
-   * Should the menu display Header Title (first-tier).
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue)
+   * Toggle for the menu display Header Title (first-tier).
+   * The default value will be updated to `false` in the next major release.
    */
   showHeader: PropTypes.bool,
 };